+++ src/flownet/network_model/_generate_connections.py
@@ -4,51 +4,57 @@ from typing import List, Tuple, Any, Optional
 from operator import itemgetter
 
 import numpy as np
+from numpy.core.function_base import linspace
 import pandas as pd
 from scipy.spatial import Delaunay, distance  # pylint: disable=no-name-in-module
 
 from ._mitchell import mitchell_best_candidate_modified_3d
+from ._hull import check_in_hull
 from ..utils.types import Coordinate
 
 
-def _create_record(
-    well_pairs: np.ndarray, dist_matrix: np.ndarray
-) -> List[Tuple[Any, ...]]:
-    """
-    Creates a record of every well_pair triangle for which one of the angles is too large
-    (currently set to be above 150 degrees)
+def _is_angle_too_large(
+    angle_threshold: float, side_a: float, side_b: float, side_c: float
+) -> bool:
+    """Function checks if there is an angle larger than a specified angle in
+    degrees for a triangle with given side lengths
 
     Args:
-        well_pairs: All well pairs created through Delaunay triangulation
-        dist_matrix: Euclidean distance between well coordinates
+        angle_threshold: threshold angle in degrees
+        side_a: Length of side a
+        side_b: Length of side b
+        side_c: Length of side c
 
     Returns:
-         A record of connections (from_well, to_well) of undesirables.
+        True if an angle larger than the specified angle
+
+    """
+    calculate_angle = np.rad2deg(
+        math.acos(
+            (math.pow(side_a, 2) + math.pow(side_b, 2) - math.pow(side_c, 2))
+            / (2 * side_a * side_b)
+        )
+    )
+
+    return calculate_angle > angle_threshold
+
+
+def _create_record(
+    connection_pairs: np.ndarray, dist_matrix: np.ndarray, angle_threshold: float
+) -> List[Tuple[Any, ...]]:
+    """
+    Creates a record of every connection_pairs triangle for which one of the angles is too small/large
+
+    Args:
+        connection_pairs: All connection_pairs created through Delaunay triangulation
+        dist_matrix: Euclidean distance between coordinate pairs
+        angle_threshold: angle in Delaunay triangles for which to ignore connections
+
+    Returns:
+         A record of undesirable connection_pairs.
 
     """
     record: List[Tuple[Any, ...]] = []
-
-    def is_angle_too_large(side_a: float, side_b: float, side_c: float) -> bool:
-        """Function checks if there is an angle larger than 150 degrees for a triangle with given side lengths
-
-        Args:
-            side_a: Length of side a
-            side_b: Length of side b
-            side_c: Length of side c
-
-        Returns:
-            True if an angle larger than 150 degrees exists
-
-        """
-        return (
-            np.rad2deg(
-                math.acos(
-                    (math.pow(side_a, 2) + math.pow(side_b, 2) - math.pow(side_c, 2))
-                    / (2 * side_a * side_b)
-                )
-            )
-            > 150
-        )
 
     for vertex_a in range(0, 2):
         for vertex_b in range(vertex_a + 1, 3):
@@ -56,14 +62,14 @@ def _create_record(
                 edge_a = dist_matrix[vertex_a, vertex_b]
                 edge_b = dist_matrix[vertex_a, vertex_c]
                 edge_c = dist_matrix[vertex_b, vertex_c]
-                if is_angle_too_large(edge_a, edge_b, edge_c):
-                    record.append(tuple(well_pairs[vertex_b, vertex_c]))
+                if _is_angle_too_large(angle_threshold, edge_a, edge_b, edge_c):
+                    record.append(tuple(connection_pairs[vertex_b, vertex_c]))
 
-                if is_angle_too_large(edge_c, edge_b, edge_a):
-                    record.append(tuple(well_pairs[vertex_a, vertex_b]))
+                if _is_angle_too_large(angle_threshold, edge_c, edge_b, edge_a):
+                    record.append(tuple(connection_pairs[vertex_a, vertex_b]))
 
-                if is_angle_too_large(edge_a, edge_c, edge_b):
-                    record.append(tuple(well_pairs[vertex_a, vertex_c]))
+                if _is_angle_too_large(angle_threshold, edge_a, edge_c, edge_b):
+                    record.append(tuple(connection_pairs[vertex_a, vertex_c]))
     return record
 
 
@@ -187,9 +193,13 @@ def _generate_connections(
                 conn_matrix[flow_node_a, flow_node_b] = conn_matrix[
                     flow_node_b, flow_node_a
                 ] = 1
-        conn_matrix = _remove_recorded_connections(
-            _create_record(well_pairs, dist_matrix), conn_matrix
-        )
+        if configuration.flownet.angle_threshold:
+            conn_matrix = _remove_recorded_connections(
+                _create_record(
+                    well_pairs, dist_matrix, configuration.flownet.angle_threshold
+                ),
+                conn_matrix,
+            )
     print("done.")
 
     # Are there nodes in the connection matrix that has no connections? Definitely a problem
@@ -300,7 +310,7 @@ def __get_entity_str(df_coordinates: pd.DataFrame, coordinate: Coordinate) -> st
     return entity
 
 
-# pylint: disable=too-many-arguments
+# pylint: disable=too-many-arguments,too-many-locals
 def _create_entity_connection_matrix(
     df_coordinates: pd.DataFrame,
     starts: List[Coordinate],
@@ -309,6 +319,8 @@ def _create_entity_connection_matrix(
     aquifer_ends: List[Coordinate],
     max_distance_fraction: float,
     max_distance: float,
+    concave_hull_bounding_boxes: Optional[np.ndarray] = None,
+    n_non_reservoir_evaluation: Optional[int] = 10,
 ) -> pd.DataFrame:
     """
     Converts the the coordinates given for starts and ends to the desired DataFrame format for simulation input.
@@ -321,6 +333,8 @@ def _create_entity_connection_matrix(
         aquifer_ends: List of coordinates of all aquifer ends
         max_distance_fraction: Fraction of longest connection distance to be removed
         max_distance: Maximum distance between nodes, removed otherwise
+        concave_hull_bounding_boxes: Numpy array with x, y, z min/max boundingboxes for each grid block
+        n_non_reservoir_evaluation: Number of equally spaced points along a connection to check fornon-reservoir.
 
     Returns:
         Connection coordinate DataFrame on Flow desired format.
@@ -342,6 +356,19 @@ def _create_entity_connection_matrix(
     for start, end in zip(starts, ends):
         str_start_entity = __get_entity_str(df_coordinates, start)
         str_end_entity = __get_entity_str(df_coordinates, end)
+
+        if concave_hull_bounding_boxes is not None:
+            tube_coordinates = linspace(
+                start=start,
+                stop=end,
+                num=n_non_reservoir_evaluation,
+                endpoint=False,
+                dtype=float,
+                axis=1,
+            ).T
+
+            if not all(check_in_hull(concave_hull_bounding_boxes, tube_coordinates)):
+                continue
 
         df_out = df_out.append(
             {
@@ -539,4 +566,6 @@ def create_connections(
         aquifer_ends,
         configuration.flownet.max_distance_fraction,
         configuration.flownet.max_distance,
+        concave_hull_bounding_boxes=concave_hull_bounding_boxes,
+        n_non_reservoir_evaluation=configuration.flownet.n_non_reservoir_evaluation,
     ).reset_index()